UI refactor complete
And the basic scene switcher functionality is live. To create a new level, open the LevelTemplate and then save as a new level under the Levels folder. Set the ZoneChanger to have a destination scene and create a new ZoneChanger in a previous scene with a path to the new destination scene.
Inventory is shared across level switches since it is decoupled now.